+++ dora/conf.py
@@ -70,6 +70,8 @@ class SlurmConfig:
             per node, otherwise, will schedule one task per gpu (default is False).
         array_parallelism (int): when using job arrays, how many tasks can run
             in parallel.
+        qos: (str or None): qos param for slurm.
+        account: (str or None): account param for slurm.
 
     ..warning:: this assumes one task per GPU.
         Set `one_task_per_node` if you do not want that.
@@ -87,6 +89,9 @@ class SlurmConfig:
     constraint: str = ""
     one_task_per_node: bool = False
     array_parallelism: int = 256
+    exclude: tp.Optional[str] = None
+    qos: tp.Optional[str] = None
+    account: tp.Optional[str] = None
 
 
 @dataclass
@@ -141,11 +146,14 @@ class DoraConfig:
         shared (Path or None): if provided, the path to a central repository of XPs.
             For the moment, this only supports sharing hyper-params, logs etc. will stay
             in the per user folder.
+        grid_package (str or None): if provided, package to look for grids. Default
+            to the package with the `train.py` module followed by `.grids`.
     """
     dir: Path = Path("./outputs")  # where everything will be stored
     exclude: tp.List[str] = field(default_factory=list)
     git_save: bool = False
     shared: tp.Optional[Path] = None  # Optional path for shared XPs.
+    grid_package: tp.Optional[str] = None
 
     # Those are internal config values and are unlikely to be changed
     history: str = "history.json"  # where metrics will be stored
@@ -154,8 +162,9 @@ class DoraConfig:
     shep: ShepConfig = field(default_factory=ShepConfig)
     rendezvous_file: str = "rendezvous.txt"
     use_rendezvous: bool = False
-    grids: str = "grids"
-    codes: str = "codes"
+    # Filenames used in various places, you shouldn't edit that
+    _grids: str = "grids"
+    _codes: str = "codes"
 
     def is_excluded(self, arg_name: str) -> bool:
         """Return True if the given argument name should be excluded from