+++ dora/distrib.py
@@ -42,7 +42,15 @@ def set_distrib_env():
         xp = get_xp()
         # Note that running twice the same XP on the same node will crash,
         # but that shouldn't really happen
-        rng = random.Random(int(xp.sig, 16))
+        seed = xp.sig
+        # If we are in a Slurm job, let us use the Slurm job id.
+        try:
+            env = submitit.JobEnvironment()
+        except RuntimeError:
+            pass
+        else:
+            seed += env.job_id
+        rng = random.Random(seed)
         master_port = rng.randint(20000, 60000)
         os.environ['MASTER_PORT'] = str(master_port)
     if 'WORLD_SIZE' not in os.environ:
@@ -56,30 +64,33 @@ def get_distrib_spec():
     This can be used even before distributed training is initialized, which is useful for
     PytorchLightning for instance.
     """
-    try:
-        env = submitit.JobEnvironment()
-    except RuntimeError:
-        if 'WORLD_SIZE' in os.environ:
-            rank = int(os.environ['RANK'])
-            world_size = int(os.environ['WORLD_SIZE'])
-            local_rank = rank
-            node_rank = 0
-            num_nodes = 1
-            source = "env"
+    if 'WORLD_SIZE' in os.environ:
+        rank = int(os.environ['RANK'])
+        world_size = int(os.environ['WORLD_SIZE'])
+        if 'LOCAL_RANK' in os.environ:
+            local_rank = int(os.environ['LOCAL_RANK'])
         else:
+            local_rank = rank
+        node_rank = 0
+        num_nodes = 1
+        source = "env"
+    else:
+        try:
+            env = submitit.JobEnvironment()
+        except RuntimeError:
             rank = 0
             world_size = 1
             local_rank = 0
             node_rank = 0
             num_nodes = 1
             source = "empty"
-    else:
-        rank = env.global_rank
-        world_size = env.num_tasks
-        local_rank = env.local_rank
-        node_rank = env.node
-        num_nodes = env.num_nodes
-        source = "submitit"
+        else:
+            rank = env.global_rank
+            world_size = env.num_tasks
+            local_rank = env.local_rank
+            node_rank = env.node
+            num_nodes = env.num_nodes
+            source = "submitit"
     return DistribSpec(rank, world_size, local_rank, node_rank, num_nodes, source)
 
 
@@ -94,7 +105,11 @@ def init(backend='nccl'):
         logger.info("world_size is 1, skipping init.")
         return
     xp = get_xp()
-    torch.cuda.set_device(spec.local_rank)
+    if torch.cuda.is_available():
+        torch.cuda.set_device(spec.local_rank)
+    else:
+        assert backend != 'nccl'
+
     if xp.dora.use_rendezvous:
         init_method = 'file://' + os.path.abspath(xp.rendezvous_file)
     else: