+++ examples/mnist_hydra/grids/test.py
@@ -22,6 +22,16 @@ def explorer(launcher):
     for bs in [32, 64, 128]:
         launcher(batch_size=bs)
 
+    for hidden_dim in [512, 1024]:
+        # here we get a sub launcher with `bind()`. All XPs scheduled with it
+        # will retain the bound params but it won't impact the parent launcher.
+        sub = launcher.bind({"model.hidden_dim": hidden_dim})
+        # Or, the two are equivalent
+        # sub = launcher.bind([f"model.hidden_dim={hidden_dim}"])
+        sub()
+        sub(gamma=0.6)
+        sub({'+new_param': 'whatever'})  # you can define extra keys with '+' if required
+
     launcher.bind_(gamma=0.6)
     launcher.slurm_(mem_per_gpu=20)
     launcher()